fixes aske by lucas
