+++ micro-services/back-office-service/src/models/ContactMessage.ts
@@ -1,20 +1,11 @@
-import { DataTypes, Model, Optional } from 'sequelize';
+import { DataTypes, Model } from 'sequelize';
 import { sequelize } from '../config/database';
+import type {
+  ContactMessageAttributes,
+  ContactMessageCreationAttributes,
+} from '../interfaces/IContactMessage';
 
-interface ContactMessageAttributes {
-  id: string;
-  email: string;
-  subject: string;
-  message: string;
-  status: 'new' | 'processed';
-  admin_reply?: string | null;
-  replied_at?: Date | null;
-  created_at?: Date;
-  updated_at?: Date;
-}
-
-export interface ContactMessageCreationAttributes
-  extends Optional<ContactMessageAttributes, 'id' | 'status' | 'admin_reply' | 'replied_at' | 'created_at' | 'updated_at'> {}
+export type { ContactMessageCreationAttributes };
 
 class ContactMessage
   extends Model<ContactMessageAttributes, ContactMessageCreationAttributes>
